requirements.txt
mysql-connector-python==9.4.0
bcrypt==4.3.0
pandas==2.2.3
colorama==0.4.6
python-dotenv==1.1.1
pytest==8.4.1
